+++ tools/matc/CMakeLists.txt
@@ -12,6 +12,7 @@ set(TARGET matlang)
 set(HDRS
         src/matc/CommandlineConfig.h
         src/matc/Compiler.h
+        src/matc/DirIncluder.h
         src/matc/MaterialCompiler.h
         )
 
@@ -19,6 +20,7 @@ set(SRCS
         src/matc/Compiler.cpp
         src/matc/MaterialCompiler.cpp
         src/matc/CommandlineConfig.cpp
+        src/matc/DirIncluder.cpp
         )
 
 # ==================================================================================================
@@ -62,3 +64,16 @@ set_target_properties(${TARGET} PROPERTIES FOLDER Tools)
 # Installation
 # ==================================================================================================
 install(TARGETS ${TARGET} RUNTIME DESTINATION bin)
+
+# ==================================================================================================
+# Tests
+# ==================================================================================================
+project(test_matc)
+set(TARGET test_matc)
+set(SRCS tests/test_includer.cpp)
+
+add_executable(${TARGET} ${SRCS})
+
+target_link_libraries(${TARGET} matlang gtest)
+
+set_target_properties(test_matc PROPERTIES FOLDER Tests)
