+++ tools/matc/tests/test_includer.cpp
@@ -16,7 +16,7 @@
 
 #include <gtest/gtest.h>
 
-#include "DirIncluder.h"
+#include <matc/DirIncluder.h>
 
 using namespace utils;
 using namespace filamat;
@@ -26,7 +26,7 @@ const utils::Path root = utils::Path(__FILE__).getParent();
 // TODO: these tests are disabled as they fail on CI, which needs investigation.
 
 TEST(DirIncluder, DISABLED_IncludeNonexistent) {
-    matp::DirIncluder includer;
+    matc::DirIncluder includer;
     {
         IncludeResult i {
             .includeName = CString("nonexistent.h")
@@ -45,7 +45,7 @@ TEST(DirIncluder, DISABLED_IncludeNonexistent) {
 }
 
 TEST(DirIncluder, DISABLED_IncludeFile) {
-    matp::DirIncluder includer;
+    matc::DirIncluder includer;
     includer.setIncludeDirectory(root);
 
     IncludeResult result {
@@ -63,7 +63,7 @@ TEST(DirIncluder, DISABLED_IncludeFile) {
 }
 
 TEST(DirIncluder, DISABLED_IncludeFileFromIncluder) {
-    matp::DirIncluder includer;
+    matc::DirIncluder includer;
     includer.setIncludeDirectory(root);
 
     utils::CString includerFile((root + "Dir/Baz.h").c_str());
@@ -76,4 +76,9 @@ TEST(DirIncluder, DISABLED_IncludeFileFromIncluder) {
     EXPECT_STREQ("// Bar.h", result.text.c_str());
 
     EXPECT_STREQ((root + "Dir/Bar.h").c_str(), result.name.c_str());
+}
+
+int main(int argc, char** argv) {
+    ::testing::InitGoogleTest(&argc, argv);
+    return RUN_ALL_TESTS();
 }
